customer-convos/requirements.txt
google-auth
google-api-python-client
simple-salesforce
sqlalchemy
psycopg2-binary
requests
